fix: .js -> .jsx로 변경
